backend/requirements.txt
Django==4.2.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
django-filter==24.1
psycopg2-binary==2.9.10
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
daphne==4.0.0
gunicorn==21.2.0
whitenoise==6.6.0
